requirements.txt
Django>=4.2,<5.0
openai>=1.30
python-dotenv>=1.0
